Improve vm script, automate docker install
The script will now fail if the subnet id isn't found rather than
continuing to attempt VM creation.

The cloud-init example is expanded to add the docker repo and install
docker as well as docker-compose.

I've added the key explicitly here, it makes the file a bit bigger but
avoids hitting any problems with poisoned keyservers.

It also mounts the ephemeral disk to `/var/lib/docker` where it will
be more useful for us.